- Bug fix related to double counting parent folder's size and the size of the child folders and files - Small refactoring. Extracted some logic into folder_size_calculator.rb concern - Small updates in the view

+++ app/controllers/folders_controller.rb
@@ -12,9 +12,10 @@ class FoldersController < ApplicationController
 
   def list_all_folders
     @folders = Folder.order(created_at: :desc).paginate(page: params[:page], per_page: 50)
-    @folder_sizes = calculate_folder_sizes(@folders)
-    @total_size = calculate_total_size(@folders)
+    @folder_sizes = calculate_total_folder_sizes(@folders)
+    @total_size = number_to_human_size(@folder_sizes.values.sum)
   end
+
 
   def new
     @folder = current_user.folders.new
@@ -52,7 +53,6 @@ class FoldersController < ApplicationController
 
     Rails.logger.info "Attempting to delete folder with ID: #{@folder.id}"
 
-    # Изтриване на всички дъщерни папки, за да се избегне циклична зависимост
     begin
       delete_descendants(@folder)
     rescue => e
@@ -61,7 +61,6 @@ class FoldersController < ApplicationController
       return
     end
 
-    # Опитваме се да изтрием физическата папка
     begin
       delete_physical_folder(@folder)
     rescue => e
@@ -70,7 +69,6 @@ class FoldersController < ApplicationController
       return
     end
 
-    # Изтриване на самата папка от базата данни
     if @folder.destroy
       Rails.logger.info "Folder successfully deleted from the database with ID: #{@folder.id}"
       redirect_to folders_path, notice: 'Folder deleted successfully.'
@@ -106,59 +104,37 @@ class FoldersController < ApplicationController
     end
   end
 
-
   def rename_physical_folder_for_user(folder)
     root_path = Rails.configuration.user_files_path
     user_folder_path = File.join(root_path, current_user.email)
 
-    # Построяване на стария и новия път
     old_folder_path = File.join(user_folder_path, build_path_for_parent(folder_was_parent(folder)), folder.name_was)
     new_folder_path = File.join(user_folder_path, build_path_for_parent(Folder.find_by(id: folder_params[:parent_id])), folder_params[:name])
 
-    Rails.logger.info "Attempting to move folder from #{old_folder_path} to #{new_folder_path}"
+    return if old_folder_path == new_folder_path
 
-    # Ако старият път е същият като новия, няма нужда от преместване
-    if old_folder_path == new_folder_path
-      Rails.logger.info "No need to move folder. The path is the same."
-      return
-    end
-
-    # Уверяване, че новата директория съществува
     FileUtils.mkdir_p(File.dirname(new_folder_path)) unless Dir.exist?(File.dirname(new_folder_path))
 
     begin
-      # Преместване на папката
-      if Dir.exist?(old_folder_path)
-        FileUtils.mv(old_folder_path, new_folder_path)
-        Rails.logger.info "Successfully moved folder from #{old_folder_path} to #{new_folder_path}"
-      else
-        Rails.logger.warn "Old folder path does not exist: #{old_folder_path}. Cannot move folder."
-      end
+      FileUtils.mv(old_folder_path, new_folder_path) if Dir.exist?(old_folder_path)
     rescue => e
-      Rails.logger.error "Error moving folder from #{old_folder_path} to #{new_folder_path}: #{e.message}"
+      Rails.logger.error "Error moving folder: #{e.message}"
       raise
     end
   end
-
-
 
   def folder_was_parent(folder)
     Folder.find_by(id: folder.parent_id_was)
   end
 
   def build_path_for_parent(parent_folder)
-    if parent_folder
-      File.join(build_path_for_parent(parent_folder.parent), parent_folder.name)
-    else
-      ''
-    end
+    parent_folder ? File.join(build_path_for_parent(parent_folder.parent), parent_folder.name) : ''
   end
-
 
   def delete_descendants(folder)
     folder.subfolders.each do |subfolder|
       delete_descendants(subfolder)
-      delete_physical_folder(subfolder) # Уверяваме се, че физическата папка също се изтрива
+      delete_physical_folder(subfolder)
       subfolder.destroy
     end
   end
@@ -168,19 +144,6 @@ class FoldersController < ApplicationController
     user_folder_path = File.join(root_path, folder.user.email)
     folder_path = File.join(user_folder_path, build_path_for_parent(folder.parent), folder.name)
 
-    if Dir.exist?(folder_path)
-      begin
-        Rails.logger.info "Deleting folder at path: #{folder_path}"
-        FileUtils.rm_rf(folder_path)
-        Rails.logger.info "Successfully deleted folder at path: #{folder_path}"
-      rescue => e
-        Rails.logger.error "Failed to delete folder at #{folder_path}: #{e.message}"
-        raise
-      end
-    else
-      Rails.logger.warn "Folder at #{folder_path} does not exist, skipping physical deletion."
-    end
+    FileUtils.rm_rf(folder_path) if Dir.exist?(folder_path)
   end
-
-
 end
